Refactoring the Navbar and Sidebar components.
