+++ data/menu-items.ts
@@ -1,0 +1,16 @@
+const menuItems = [
+  {
+    title: 'Board',
+    icon: 'PiBriefcaseLight',
+  },
+  {
+    title: 'Contacts',
+    icon: 'RiContactsLine',
+  },
+  {
+    title: 'Documents',
+    icon: 'RiFolder2Line',
+  },
+];
+
+export default menuItems;
