+++ types/index.d.ts
@@ -1,0 +1,4 @@
+type MenuItemProps = {
+  linkName: string;
+  icon: JSX.Element;
+};
